+++ NEWS.md
@@ -1,10 +1,14 @@
 MathOptInterface (MOI) release notes
 ====================================
 
-v0.9.10 (January 31, 2020)
+v0.9.11 (February 9, 2020)
 ---------------------
 
 - Added `GeoMeantoRelEntrBridge` to bridge a geomean constraint to a relative entropy constraint.
+
+v0.9.10 (January 31, 2020)
+---------------------
+
 - Added `OptimizerWithAttributes` grouping an optimizer constructor and a list
   of optimizer attributes (#1008).
 - Added `RelativeEntropyCone` with corresponding bridge into exponential cone